pyproject.toml
[project]
name = "hrms-lite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.1",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.29.0",
    "markdown>=3.10.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
]
